# backend/internal/feature/playback/stream_handlers.go
package playback

import (
	"context"
	"errors"
	"fmt"
	"io"
	"net/http"
	"os"
	"os/exec"
	"path/filepath"
	"regexp"
	"strconv"
	"strings"

	"github.com/google/uuid"

	"github.com/go-chi/chi/v5"

	"couchverse/internal/feature/auth"
	"couchverse/internal/feature/catalog"
	"couchverse/internal/feature/jobs"
	"couchverse/internal/feature/library"
	"couchverse/internal/feature/subtitles"
	"couchverse/internal/grant"
	"couchverse/internal/httpx"
	"couchverse/internal/media"
	"couchverse/internal/settings"
)

type Stream struct {
	subs     *subtitles.Store
	catalog  *catalog.Store
	library  *library.Store
	settings *settings.Store
	jobs     *jobs.Store
	dataDir  string
	sessions *SessionManager
	ffmpeg   string
	grants   *grant.Signer
}

func NewStream(subs *subtitles.Store, cat *catalog.Store, lib *library.Store, set *settings.Store, jb *jobs.Store, dataDir string, sessions *SessionManager, ffmpegPath string, grants *grant.Signer) *Stream {
	return &Stream{subs: subs, catalog: cat, library: lib, settings: set, jobs: jb, dataDir: dataDir, sessions: sessions, ffmpeg: ffmpegPath, grants: grants}
}

// MediaPath is the URL of a route under a media grant: /api/v1/media/{grant}/<rest>.
func MediaPath(g, rest string) string {
	return "/api/v1/media/" + g + "/" + rest
}

// grantedFile is the media file the request's grant names (checked by the media
// group's middleware).
func grantedFile(ctx context.Context) string {
	g, _ := grant.From(ctx)
	return g.Resource.String()
}

// Viewer is who a playback payload is built for.
type Viewer struct {
	// UserID is the signed-in viewer, 0 for an anonymous couch guest; it gets
	// the saved resume position.
	UserID int64
	// Couch is the couch participant id of a follower, "" otherwise; it binds
	// the follower's grants to their place on the couch.
	Couch string
}

// mediaGrant signs a grant for one media file to the viewer.
func (h *Stream) mediaGrant(mediaFileID string, v Viewer) string {
	g := grant.Grant{Scope: grant.Media, Subject: v.UserID}
	var err error
	if g.Resource, err = uuid.Parse(mediaFileID); err != nil {
		return ""
	}
	if v.Couch != "" {
		if g.Couch, err = uuid.Parse(v.Couch); err != nil {
			return ""
		}
	}
	return h.grants.Issue(g, grant.MediaTTL)
}

var contentTypes = map[string]string{
	"mp4":  "video/mp4",
	"m4v":  "video/mp4",
	"webm": "video/webm",
	"mkv":  "video/x-matroska",
	"mp3":  "audio/mpeg",
	"flac": "audio/flac",
	"m4a":  "audio/mp4",
	"ogg":  "audio/ogg",
	"opus": "audio/ogg",
	"wav":  "audio/wav",
}

// Frame returns a single still from the source video at ?t=SECONDS, scaled
// down for the player's seek-bar preview. Requests are bucketed to a few
// seconds and cached so scrubbing reuses extracted frames.
func (h *Stream) Frame(w http.ResponseWriter, r *http.Request) {
	mf, err := h.library.MediaFileByID(r.Context(), grantedFile(r.Context()))
	if err != nil {
		httpx.StoreErr(w, err)
		return
	}
	if mf.SourceDeletedAt != nil || mf.VideoCodec == "" {
		httpx.NotFound(w) // no source on disk to grab a frame from
		return
	}

	const bucket = 5 // seconds; coarse enough to bound the cache and reuse hovers
	t := httpx.QueryInt(r, "t", 0)
	if t < 0 {
		t = 0
	}
	if mf.DurationSeconds > 0 && float64(t) > mf.DurationSeconds {
		t = int(mf.DurationSeconds)
	}
	t = (t / bucket) * bucket

	dir := filepath.Join(h.dataDir, "cache", "frames", mf.ID)
	cached := filepath.Join(dir, strconv.Itoa(t)+".jpg")
	if _, err := os.Stat(cached); err != nil {
		lib, lerr := h.library.LibraryByID(r.Context(), mf.LibraryID)
		if lerr != nil {
			httpx.Internal(w, lerr)
			return
		}
		if err := os.MkdirAll(dir, 0o755); err != nil {
			httpx.Internal(w, err)
			return
		}
		// -ss before -i is a fast input seek to the nearest keyframe
		cmd := exec.CommandContext(r.Context(), h.ffmpeg,
			"-hide_banner", "-loglevel", "error", "-y",
			"-ss", strconv.Itoa(t), "-i", filepath.Join(lib.Path, mf.Path),
			"-frames:v", "1", "-vf", "scale=240:-2", "-q:v", "5", cached)
		if out, err := cmd.CombinedOutput(); err != nil {
			os.Remove(cached)
			httpx.Error(w, http.StatusNotFound, "frame_failed", strings.TrimSpace(string(out)))
			return
		}
	}
	w.Header().Set("Cache-Control", "public, max-age=86400")
	httpx.ServeFile(w, r, cached)
}

// Serve streams a media file with HTTP range support (direct play).
func (h *Stream) Serve(w http.ResponseWriter, r *http.Request) {
	mf, err := h.library.MediaFileByID(r.Context(), grantedFile(r.Context()))
	if err != nil {
		httpx.StoreErr(w, err)
		return
	}
	if mf.SourceDeletedAt != nil {
		httpx.Error(w, http.StatusNotFound, "source_deleted", "the original file was removed after transcoding")
		return
	}
	lib, err := h.library.LibraryByID(r.Context(), mf.LibraryID)
	if err != nil {
		httpx.Internal(w, err)
		return
	}

	f, err := os.Open(filepath.Join(lib.Path, mf.Path))
	if err != nil {
		httpx.Error(w, http.StatusNotFound, "file_missing", "media file is missing on disk")
		return
	}
	defer f.Close()

	info, err := f.Stat()
	if err != nil {
		httpx.Internal(w, err)
		return
	}

	if ct, ok := contentTypes[mf.Container]; ok {
		w.Header().Set("Content-Type", ct)
	}
	http.ServeContent(w, r, filepath.Base(mf.Path), info.ModTime(), f)
}

// PlaybackInfo is the player payload. It is built by BuildPlayback and reused
// by the couch feature to assemble a follower's player without an auth context.
type PlaybackInfo struct {
	Mode           string                  `json:"mode" enum:"direct,hls,preparing,jit,unsupported" doc:"How to play: direct and hls load streamUrl; preparing waits for a running transcode (see jobProgress); jit opens a session with createStreamSession; unsupported cannot play."`
	MediaFileID    string                  `json:"mediaFileId"`
	Grant          string                  `json:"grant" doc:"The media grant every URL in this payload carries; it expires, so fetch the payload again on grant_expired."`
	StreamURL      string                  `json:"streamUrl,omitempty" doc:"The source file (mode direct) or the HLS master (mode hls) to load."`
	FrameURL       string                  `json:"frameUrl" doc:"A still for the seek-bar preview; append ?t=<seconds>."`
	Duration       float64                 `json:"durationSeconds"`
	ResumePosition int                     `json:"resumePosition" doc:"Saved position in seconds; 0 for couch followers, who sync to the host."`
	Display        PlaybackDisplay         `json:"display"`
	NextEpisode    *catalog.EpisodeRef     `json:"nextEpisode,omitempty" doc:"The episode after this one; absent for movies and series finales."`
	Subtitles      []PlaybackSubtitleTrack `json:"subtitles"`
	Audio          []PlaybackAudioTrack    `json:"audio,omitempty" doc:"Selectable audio languages; absent when there is only one."`
	Episodes       []catalog.SeriesEpisode `json:"episodes,omitempty" doc:"The series' playable episodes, for the in-player switcher."`
	CurrentEpisode string                  `json:"currentEpisodeId,omitempty"`
	HLSURL         string                  `json:"hlsUrl,omitempty" doc:"HLS master of the ready transcodes, offered even when the source direct-plays."`
	Variants       []QualityVariant        `json:"variants,omitempty" doc:"Ready renditions for the quality menu."`
	JobProgress    int                     `json:"jobProgress,omitempty" doc:"Transcode progress in percent while mode is preparing."`
	// series opt-in for shuffle playback (drives the player's shuffle toggle)
	AllowRandomPlayback bool `json:"allowRandomPlayback"`
}

// QualityVariant is one ready rendition in the player's quality menu.
type QualityVariant struct {
	Name   string `json:"name"`
	Height int    `json:"height"`
}

type PlaybackSubtitleTrack struct {
	ID     string `json:"id"`
	Lang   string `json:"lang"`
	Label  string `json:"label"`
	Forced bool   `json:"forced"`
	URL    string `json:"url" doc:"The track as WebVTT."`
}

// PlaybackAudioTrack is one selectable audio language. Source "file" (model B)
// is a separate-language media file the player swaps to; "embedded" (model A)
// is an in-stream HLS audio rendition.
type PlaybackAudioTrack struct {
	ID        string `json:"id" doc:"The media file id of a file track; embedded:<stream index> for an embedded one."`
	Lang      string `json:"lang"`
	Label     string `json:"label"`
	Default   bool   `json:"default"`
	Source    string `json:"source" enum:"file,embedded"`
	StreamURL string `json:"streamUrl,omitempty" doc:"Direct stream of a file track that direct-plays."`
	HLSURL    string `json:"hlsUrl,omitempty" doc:"HLS master of a file track that does not direct-play."`
}

var audioLangNames = map[string]string{
	"en": "English", "cs": "Čeština", "sk": "Slovenčina", "de": "Deutsch",
	"es": "Español", "fr": "Français", "it": "Italiano", "pl": "Polski",
	"ko": "한국어", "ja": "日本語", "ru": "Русский", "zh": "中文",
}

func audioLabel(lang string) string {
	if lang == "" || lang == "und" {
		return "Original"
	}
	if n, ok := audioLangNames[lang]; ok {
		return n
	}
	return strings.ToUpper(lang)
}

// audioTrackFor builds a model-B track from a media file, pointing at its direct
// stream when it direct-plays, or its HLS master otherwise. Each file needs its
// own grant.
func audioTrackFor(mf *media.MediaFile, g string, isDefault bool) PlaybackAudioTrack {
	t := PlaybackAudioTrack{ID: mf.ID, Lang: mf.AudioLang, Label: audioLabel(mf.AudioLang), Default: isDefault, Source: "file"}
	if mf.SourceDeletedAt == nil && mf.DirectPlay {
		t.StreamURL = MediaPath(g, "stream")
	} else {
		t.HLSURL = MediaPath(g, "hls/master.m3u8")
	}
	return t
}

// PlaybackDisplay is what the player shows about the title being played.
type PlaybackDisplay struct {
	Title          string  `json:"title"`
	Subtitle       string  `json:"subtitle" doc:"The episode line (season, episode and name); empty for movies."`
	TitleID        string  `json:"titleId"`
	TitleSlug      string  `json:"titleSlug"`
	BackdropID     *string `json:"backdropId"`
	BackdropVer    int64   `json:"backdropVer,omitempty"`
	BackdropAccent string  `json:"backdropAccent,omitempty"`
}

var (
	errNoMedia = errors.New("playback: title or episode has no media file")
	errBadKind = errors.New("playback: kind must be movie or episode")
)

type playbackInput struct {
	Kind string   `path:"kind" enum:"movie,episode"`
	ID   string   `path:"id" format:"uuid" doc:"The movie's title id or the episode id."`
	Caps []string `query:"caps" doc:"Video codecs the client decodes beyond the h264/vp9/av1 baseline (e.g. hevc), for the direct-play decision."`
}

type playbackOutput struct{ Body *PlaybackInfo }

// Playback resolves what to play for a movie title or an episode.
func (h *Stream) Playback(ctx context.Context, in *playbackInput) (*playbackOutput, error) {
	info, err := h.BuildPlayback(ctx, in.Kind, in.ID, Viewer{UserID: auth.UserFrom(ctx).ID}, in.Caps)
	if errors.Is(err, errNoMedia) {
		return nil, httpx.Fail(http.StatusNotFound, "no_media", "this title has no media file yet")
	}
	if err != nil {
		return nil, err
	}
	return &playbackOutput{Body: info}, nil
}

// BuildPlayback assembles the player payload for a movie title or an episode,
// with media grants issued to the viewer. A viewer outside a couch gets their
// saved resume position; followers sync to the host instead. It reads no auth
// and writes no response, so the couch feature reuses it to build a follower's
// payload. caps is the client's container/codec capability list (used for the
// direct-play decision).
func (h *Stream) BuildPlayback(ctx context.Context, kind, id string, viewer Viewer, caps []string) (*PlaybackInfo, error) {
	// resume positions are personal; a couch plays wherever the host is
	userID := viewer.UserID
	if viewer.Couch != "" {
		userID = 0
	}
	if id == "" {
		return nil, httpx.ErrNotFound
	}

	var (
		mf   *media.MediaFile
		err  error
		info PlaybackInfo
	)

	switch kind {
	case "movie":
		title, terr := h.catalog.TitleByID(ctx, id)
		if terr != nil {
			return nil, terr
		}
		mf, err = h.catalog.PrimaryMediaFileForTitle(ctx, id)
		if err != nil {
			return nil, errNoMedia
		}
		info.Display = PlaybackDisplay{Title: title.Name, TitleID: title.ID, TitleSlug: title.Slug}
		info.AllowRandomPlayback = title.AllowRandomPlayback
		if userID != 0 {
			pos, _, perr := h.catalog.ProgressFor(ctx, userID, &id, nil)
			if perr != nil {
				return nil, perr
			}
			info.ResumePosition = pos
		}

	case "episode":
		ref, rerr := h.catalog.EpisodeRef(ctx, id)
		if rerr != nil {
			return nil, httpx.ErrNotFound
		}
		mf, err = h.catalog.PrimaryMediaFileForEpisode(ctx, id)
		if err != nil {
			return nil, errNoMedia
		}
		info.Display = PlaybackDisplay{
			Title:     ref.TitleName,
			Subtitle:  formatEpisodeSubtitle(ref),
			TitleID:   ref.TitleID,
			TitleSlug: ref.TitleSlug,
		}
		// the shuffle flag lives on the title; only the episode ref is loaded above
		if t, terr := h.catalog.TitleByID(ctx, ref.TitleID); terr == nil {
			info.AllowRandomPlayback = t.AllowRandomPlayback
		}
		if userID != 0 {
			pos, _, perr := h.catalog.ProgressFor(ctx, userID, nil, &id)
			if perr != nil {
				return nil, perr
			}
			info.ResumePosition = pos
		}
		if next, nerr := h.catalog.NextEpisode(ctx, id); nerr == nil {
			info.NextEpisode = next
		}
		if eps, eerr := h.catalog.PlayableEpisodes(ctx, ref.TitleID); eerr == nil {
			info.Episodes = eps
			info.CurrentEpisode = id
		}

	default:
		return nil, errBadKind
	}

	// banner accent for the player (best effort)
	if bid, bver, accent, berr := h.catalog.TitleBackdrop(ctx, info.Display.TitleID); berr == nil {
		info.Display.BackdropID = bid
		info.Display.BackdropVer = bver
		info.Display.BackdropAccent = accent
	}

	info.MediaFileID = mf.ID
	info.Duration = mf.DurationSeconds
	g := h.mediaGrant(mf.ID, viewer)
	info.Grant = g
	info.FrameURL = MediaPath(g, "frame")

	subs, err := h.subs.SubtitlesForMediaFile(ctx, mf.ID)
	if err != nil {
		return nil, err
	}
	info.Subtitles = []PlaybackSubtitleTrack{}
	for _, sub := range subs {
		info.Subtitles = append(info.Subtitles, PlaybackSubtitleTrack{
			ID:     sub.ID,
			Lang:   sub.Lang,
			Label:  sub.Label,
			Forced: sub.Forced,
			URL:    MediaPath(g, "subtitles/"+sub.ID+".vtt"),
		})
	}

	// alternate-audio siblings (model B): a language switch in the player. Only
	// populated when the title/episode has more than one audio file.
	if siblings, serr := h.catalog.AudioSiblings(ctx, mf.TitleID, mf.EpisodeID, mf.ID); serr == nil && len(siblings) > 0 {
		info.Audio = append(info.Audio, audioTrackFor(mf, g, true))
		for i := range siblings {
			info.Audio = append(info.Audio, audioTrackFor(&siblings[i], h.mediaGrant(siblings[i].ID, viewer), false))
		}
	}

	// embedded multi-audio (model A): a file with >=2 audio tracks streams via the
	// var_stream_map HLS remux so the player can switch audio language.
	audioStreams, _ := h.library.AudioStreamsForFile(ctx, mf.ID)
	multiAudio := len(audioStreams) >= 2

	// ready transcode variants power the player's quality menu and are offered
	// even when the source direct-plays, so users can pick a specific rendition
	variants, verr := h.library.VariantsForMediaFile(ctx, mf.ID)
	if verr != nil {
		return nil, verr
	}
	ready, pending, multiAudioReady := false, false, false
	for _, v := range variants {
		switch v.Status {
		case "ready":
			ready = true
			switch v.Name {
			case "multiaudio":
				multiAudioReady = true
			case "source": // the "source" remux isn't a distinct quality
			default:
				height := v.Height
				if v.Mode == "copy" {
					height = mf.Height
				}
				info.Variants = append(info.Variants, QualityVariant{Name: v.Name, Height: height})
			}
		case "queued", "processing":
			pending = true
		}
	}
	hlsURL := MediaPath(g, "hls/master.m3u8")
	if multiAudio && multiAudioReady {
		hlsURL = MediaPath(g, "hls/multiaudio/master.m3u8")
	}
	if ready {
		info.HLSURL = hlsURL
	}

	// embedded audio tracks for the player's language menu (model A)
	if multiAudio && len(info.Audio) == 0 {
		for _, a := range audioStreams {
			label := a.Title
			if label == "" || label == a.Lang {
				label = audioLabel(a.Lang)
			}
			info.Audio = append(info.Audio, PlaybackAudioTrack{
				ID:      fmt.Sprintf("embedded:%d", a.Index),
				Lang:    a.Lang,
				Label:   label,
				Default: a.Default,
				Source:  "embedded",
			})
		}
		hasDef := false
		for _, t := range info.Audio {
			if t.Default {
				hasDef = true
			}
		}
		if !hasDef && len(info.Audio) > 0 {
			info.Audio[0].Default = true
		}
	}

	switch {
	// embedded multi-audio must use the HLS remux so the player can switch audio,
	// even when the source would otherwise direct-play
	case multiAudio && multiAudioReady:
		info.Mode = "hls"
		info.StreamURL = hlsURL

	// a cleaned-up source can't be served directly no matter what the caps say
	case mf.SourceDeletedAt == nil && !multiAudio &&
		(mf.DirectPlay || media.DirectPlayWithCaps(mf.Container, mf.VideoCodec, mf.AudioCodec, caps)):
		info.Mode = "direct"
		info.StreamURL = MediaPath(g, "stream")

	case ready:
		info.Mode = "hls"
		info.StreamURL = hlsURL
	case pending:
		info.Mode = "preparing"
		if progress, perr := h.jobs.TranscodeProgress(ctx, mf.ID); perr == nil {
			info.JobProgress = progress
		}
	case mf.SourceDeletedAt == nil && h.jitAllowed(ctx):
		// the client opens a JIT session with createStreamSession
		info.Mode = "jit"
	default:
		info.Mode = "unsupported"
	}

	return &info, nil
}

// HLSMaster generates the master playlist from ready variants.
func (h *Stream) HLSMaster(w http.ResponseWriter, r *http.Request) {
	mediaFileID := grantedFile(r.Context())
	mf, err := h.library.MediaFileByID(r.Context(), mediaFileID)
	if err != nil {
		httpx.StoreErr(w, err)
		return
	}
	variants, err := h.library.VariantsForMediaFile(r.Context(), mediaFileID)
	if err != nil {
		httpx.Internal(w, err)
		return
	}

	var b strings.Builder
	b.WriteString("#EXTM3U\n#EXT-X-VERSION:3\n")
	count := 0
	for _, v := range variants {
		if v.Status != "ready" {
			continue
		}
		bandwidth := v.VideoBitrate + v.AudioBitrate
		if bandwidth <= 0 {
			bandwidth = mf.Bitrate
		}
		width := v.Width
		height := v.Height
		if v.Mode == "copy" {
			width, height = mf.Width, mf.Height
		}
		if width == 0 && height > 0 && mf.Height > 0 {
			width = mf.Width * height / mf.Height
		}
		fmt.Fprintf(&b, "#EXT-X-STREAM-INF:BANDWIDTH=%d,RESOLUTION=%dx%d,NAME=\"%s\"\n%s/index.m3u8\n",
			bandwidth, width, height, v.Name, v.Name)
		count++
	}
	if count == 0 {
		httpx.NotFound(w)
		return
	}
	w.Header().Set("Content-Type", "application/vnd.apple.mpegurl")
	w.Header().Set("Cache-Control", "no-cache")
	io.WriteString(w, b.String())
}

// jitAllowed: explicit setting wins; auto enables JIT when a hardware
// encoder was detected (software JIT is usually too slow for live seeking).
func (h *Stream) jitAllowed(ctx context.Context) bool {
	settings := media.LoadTranscodeSettings(ctx, h.settings)
	if settings.JITEnabled != nil {
		return *settings.JITEnabled
	}
	return len(DetectEncoders(h.ffmpeg)) > 0
}

// requireJIT gates opening a JIT session. It runs before the request is
// parsed, so a disabled instant play is reported ahead of any input error.
func (h *Stream) requireJIT(ctx context.Context) error {
	if !h.jitAllowed(ctx) {
		return httpx.Fail(http.StatusPreconditionFailed, "jit_disabled", "instant play is disabled")
	}
	return nil
}

type StreamSessionStart struct {
	StartAt float64 `json:"startAt" required:"false" doc:"Position in seconds to start transcoding from."`
}

type createStreamSessionInput struct {
	Body StreamSessionStart
}

// StreamSession is an open JIT transcode. Play PlaylistURL and keep the
// session alive while watching; idle sessions are reaped.
type StreamSession struct {
	SessionID   string `json:"sessionId"`
	PlaylistURL string `json:"playlistUrl"`
}

type streamSessionOutput struct{ Body StreamSession }

// CreateSession opens a JIT transcode session for the granted media file.
func (h *Stream) CreateSession(ctx context.Context, in *createStreamSessionInput) (*streamSessionOutput, error) {
	g, _ := grant.From(ctx)
	mediaFileID := g.Resource.String()
	if mf, merr := h.library.MediaFileByID(ctx, mediaFileID); merr == nil && mf.SourceDeletedAt != nil {
		return nil, httpx.Fail(http.StatusNotFound, "source_deleted", "the original file was removed after transcoding")
	}
	session, err := h.sessions.Create(ctx, context.Background(), mediaFileID, max(0, in.Body.StartAt))
	if err != nil {
		return nil, httpx.Fail(http.StatusServiceUnavailable, "session_failed", err.Error())
	}
	return &streamSessionOutput{Body: StreamSession{
		SessionID: session.ID,
		// the grant in the path lets relative segment URIs inherit it
		PlaylistURL: MediaPath(h.grants.Sign(g), "jit/"+session.ID+"/index.m3u8"),
	}}, nil
}

type sessionInput struct {
	SID string `path:"sid" pattern:"^[a-f0-9]{24}$"`
}

// grantedSession finds a JIT session of the granted media file; a grant for one
// file never reaches another file's session.
func (h *Stream) grantedSession(ctx context.Context, sid string) *Session {
	session := h.sessions.Get(sid)
	if session == nil || session.MediaFileID != grantedFile(ctx) {
		return nil
	}
	return session
}

func (h *Stream) SessionKeepalive(ctx context.Context, in *sessionInput) (*struct{}, error) {
	if h.grantedSession(ctx, in.SID) == nil || !h.sessions.Touch(in.SID) {
		return nil, httpx.NotFoundError()
	}
	return nil, nil
}

// StopSession ends a JIT session as soon as the player leaves, instead of
// holding the transcoder until the idle reaper notices.
func (h *Stream) StopSession(ctx context.Context, in *sessionInput) (*struct{}, error) {
	if h.grantedSession(ctx, in.SID) == nil {
		return nil, httpx.NotFoundError()
	}
	h.sessions.Stop(in.SID)
	return nil, nil
}

var sessionIDRe = regexp.MustCompile(`^[a-f0-9]{24}$`)

func (h *Stream) SessionFile(w http.ResponseWriter, r *http.Request) {
	sid := chi.URLParam(r, "sid")
	file := chi.URLParam(r, "file")
	if !sessionIDRe.MatchString(sid) {
		httpx.BadRequest(w, "invalid session id")
		return
	}
	session := h.grantedSession(r.Context(), sid)
	if session == nil {
		httpx.NotFound(w)
		return
	}

	if file == "index.m3u8" {
		w.Header().Set("Content-Type", "application/vnd.apple.mpegurl")
		w.Header().Set("Cache-Control", "no-cache")
		io.WriteString(w, session.Playlist())
		return
	}

	path, err := h.sessions.SegmentPath(r.Context(), session, file)
	if err != nil {
		httpx.Error(w, http.StatusNotFound, "segment_unavailable", err.Error())
		return
	}
	w.Header().Set("Content-Type", hlsContentType(file))
	w.Header().Set("Cache-Control", "private, max-age=60")
	httpx.ServeFile(w, r, path)
}

var hlsFileRe = regexp.MustCompile(`^[A-Za-z0-9_.-]+$`)

// HLSFile serves variant playlists and segments from the HLS cache.
func (h *Stream) HLSFile(w http.ResponseWriter, r *http.Request) {
	mediaFileID := grantedFile(r.Context())
	variant := chi.URLParam(r, "variant")
	file := chi.URLParam(r, "file")
	if !hlsFileRe.MatchString(variant) || !hlsFileRe.MatchString(file) {
		httpx.BadRequest(w, "invalid path")
		return
	}
	path := filepath.Join(h.dataDir, "cache", "hls", mediaFileID, variant, file)
	if ct := hlsContentType(file); ct != "" {
		w.Header().Set("Content-Type", ct)
	}
	w.Header().Set("Cache-Control", "private, max-age=3600")
	httpx.ServeFile(w, r, path)
}

// hlsContentType names HLS files explicitly: Go's extension table lacks .m4s
// and maps .ts to TypeScript on some systems, which strict players reject.
func hlsContentType(file string) string {
	switch filepath.Ext(file) {
	case ".m3u8":
		return "application/vnd.apple.mpegurl"
	case ".ts":
		return "video/mp2t"
	case ".m4s":
		return "video/iso.segment"
	case ".mp4":
		return "video/mp4"
	case ".vtt", ".webvtt":
		return "text/vtt"
	case ".aac":
		return "audio/aac"
	}
	return ""
}

func formatEpisodeSubtitle(ref *catalog.EpisodeRef) string {
	s := fmt.Sprintf("S%d E%d", ref.SeasonNumber, ref.EpisodeNumber)
	if ref.Name != "" {
		s += " · " + ref.Name
	}
	return s
}