# contract/openapi.json
{
  "components": {
    "schemas": {
      "APIError": {
        "additionalProperties": false,
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "required": [
          "error"
        ],
        "type": "object"
      },
...
      "CouchMediaRef": {
        "additionalProperties": false,
        "properties": {
          "episodeId": {
            "type": "string"
          },
          "kind": {
            "description": "Empty while the host is choosing what to watch.",
            "enum": [
              "",
              "movie",
              "episode"
            ],
            "type": "string"
          },
          "titleId": {
            "type": "string"
          }
        },
        "required": [
          "kind"
        ],
        "type": "object"
      },
...
      "CouchPlayback": {
        "additionalProperties": false,
        "properties": {
          "media": {
            "$ref": "#/components/schemas/CouchMediaRef"
          },
          "player": {
            "$ref": "#/components/schemas/PlaybackInfo",
            "description": "Absent while the host is choosing what to watch."
          }
        },
        "required": [
          "media"
        ],
        "type": "object"
      },
...
      "ErrorBody": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "type": "object"
      },
...
      "PlaybackInfo": {
        "additionalProperties": false,
        "properties": {
          "allowRandomPlayback": {
            "type": "boolean"
          },
          "audio": {
            "description": "Selectable audio languages; absent when there is only one.",
            "items": {
              "$ref": "#/components/schemas/PlaybackAudioTrack"
            },
            "type": "array"
          },
          "currentEpisodeId": {
            "type": "string"
          },
          "display": {
            "$ref": "#/components/schemas/PlaybackDisplay"
          },
          "durationSeconds": {
            "format": "double",
            "type": "number"
          },
          "episodes": {
            "description": "The series' playable episodes, for the in-player switcher.",
            "items": {
              "$ref": "#/components/schemas/SeriesEpisode"
            },
            "type": "array"
          },
          "frameUrl": {
            "description": "A still for the seek-bar preview; append ?t=\u003cseconds\u003e.",
            "type": "string"
          },
          "grant": {
            "description": "The media grant every URL in this payload carries; it expires, so fetch the payload again on grant_expired.",
            "type": "string"
          },
          "hlsUrl": {
            "description": "HLS master of the ready transcodes, offered even when the source direct-plays.",
            "type": "string"
          },
          "jobProgress": {
            "description": "Transcode progress in percent while mode is preparing.",
            "format": "int64",
            "type": "integer"
          },
          "mediaFileId": {
            "type": "string"
          },
          "mode": {
            "description": "How to play: direct and hls load streamUrl; preparing waits for a running transcode (see jobProgress); jit opens a session with createStreamSession; unsupported cannot play.",
            "enum": [
              "direct",
              "hls",
              "preparing",
              "jit",
              "unsupported"
            ],
            "type": "string"
          },
          "nextEpisode": {
            "$ref": "#/components/schemas/EpisodeRef",
            "description": "The episode after this one; absent for movies and series finales."
          },
          "resumePosition": {
            "description": "Saved position in seconds; 0 for couch followers, who sync to the host.",
            "format": "int64",
            "type": "integer"
          },
          "streamUrl": {
            "description": "The source file (mode direct) or the HLS master (mode hls) to load.",
            "type": "string"
          },
          "subtitles": {
            "items": {
              "$ref": "#/components/schemas/PlaybackSubtitleTrack"
            },
            "type": "array"
          },
          "variants": {
            "description": "Ready renditions for the quality menu.",
            "items": {
              "$ref": "#/components/schemas/QualityVariant"
            },
            "type": "array"
          }
        },
        "required": [
          "mode",
          "mediaFileId",
          "grant",
          "frameUrl",
          "durationSeconds",
          "resumePosition",
          "display",
          "subtitles",
          "allowRandomPlayback"
        ],
        "type": "object"
      },
...
      "StreamSession": {
        "additionalProperties": false,
        "properties": {
          "playlistUrl": {
            "type": "string"
          },
          "sessionId": {
            "type": "string"
          }
        },
        "required": [
          "sessionId",
          "playlistUrl"
        ],
        "type": "object"
      },
      "StreamSessionStart": {
        "additionalProperties": false,
        "properties": {
          "startAt": {
            "description": "Position in seconds to start transcoding from.",
            "format": "double",
            "type": "number"
          }
        },
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearerDevice": {
        "description": "Native clients: a device token from signInDevice, pollPairing or connectDevice.",
        "scheme": "bearer",
        "type": "http"
      },
      "cookieSession": {
        "description": "Browsers.",
        "in": "cookie",
        "name": "couchverse_session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "title": "Couchverse",
    "version": "1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/couch/{token}/playback": {
      "get": {
        "operationId": "getCouchPlayback",
        "parameters": [
          {
            "description": "Display language (ISO 639-1). Empty serves the base text.",
            "in": "query",
            "name": "lang",
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "The session's share code.",
            "in": "path",
            "name": "token",
            "required": true,
            "schema": {
              "description": "The session's share code.",
              "type": "string"
            }
          },
          {
            "description": "The participant cookie set by createCouch or joinCouch.",
            "in": "cookie",
            "name": "couchverse_couch",
            "schema": {
              "description": "The participant cookie set by createCouch or joinCouch.",
              "type": "string"
            }
          },
          {
            "description": "Video codecs the client decodes beyond the h264/vp9/av1 baseline (e.g. hevc), for the direct-play decision.",
            "explode": false,
            "in": "query",
            "name": "caps",
            "schema": {
              "description": "Video codecs the client decodes beyond the h264/vp9/av1 baseline (e.g. hevc), for the direct-play decision.",
              "items": {
                "type": "string"
              },
              "type": "array"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CouchPlayback"
                }
              }
            },
            "description": "OK"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            },
            "description": "Error"
          }
        },
        "tags": [
          "couch"
        ]
      }
    },
...
    "/media/{grant}/jit": {
      "post": {
        "operationId": "createStreamSession",
        "parameters": [
          {
            "description": "The media grant from the playback payload; it expires, so fetch the payload again on grant_expired.",
            "in": "path",
            "name": "grant",
            "required": true,
            "schema": {
              "pattern": "^[A-Za-z0-9_-]+$",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StreamSessionStart"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StreamSession"
                }
              }
            },
            "description": "Created"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            },
            "description": "Error"
          }
        },
        "tags": [
          "playback"
        ]
      }
    },
...
    "/playback/{kind}/{id}": {
      "get": {
        "operationId": "getPlayback",
        "parameters": [
          {
            "description": "Display language (ISO 639-1). Empty serves the base text.",
            "in": "query",
            "name": "lang",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "kind",
            "required": true,
            "schema": {
              "enum": [
                "movie",
                "episode"
              ],
              "type": "string"
            }
          },
          {
            "description": "The movie's title id or the episode id.",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "description": "The movie's title id or the episode id.",
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "description": "Video codecs the client decodes beyond the h264/vp9/av1 baseline (e.g. hevc), for the direct-play decision.",
            "explode": false,
            "in": "query",
            "name": "caps",
            "schema": {
              "description": "Video codecs the client decodes beyond the h264/vp9/av1 baseline (e.g. hevc), for the direct-play decision.",
              "items": {
                "type": "string"
              },
              "type": "array"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlaybackInfo"
                }
              }
            },
            "description": "OK"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            },
            "description": "Error"
          }
        },
        "security": [
          {
            "cookieSession": []
          },
          {
            "bearerDevice": []
          }
        ],
        "tags": [
          "playback"
        ]
      }
    },
...
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ]
}